Put instructions back on payment page
