+++ tests/pages/payment.test.js
@@ -1,5 +1,6 @@
 import { expect } from 'chai';
 import { Selector } from 'testcafe';
+import VueSelector from 'testcafe-vue-selectors';
 
 const el = Selector(selector => document.querySelector(selector));
 
@@ -54,4 +55,16 @@ test('Payment button works', async t => {
     const pagePath = await t.eval(() => window.location);
     expect(pagePath.pathname).eql('/account');
 
+});
+
+test('Instructions component occurs on payment page', async t => {
+    await t
+        .typeText('#userName', 'email@example.com')
+        .typeText('#userPassword', 'secretpass')
+        .click('button')
+        .click("#radio-0")
+        .click('button#continue')
+    ;
+    const instructionsComp = VueSelector('Instructions').exists;
+    expect(instructionsComp).to.be.ok;
 });
